Add Panic Handling to ListUsers expandIntersection
