+++ pkg/server/commands/listusers/list_users_rpc_test.go
@@ -4106,6 +4106,47 @@ func TestListUsersExclusionPanicExpandDirect(t *testing.T) {
 	tests.runListUsersTestCases(t)
 }
 
+func TestListUsersIntersectionPanic(t *testing.T) {
+	t.Cleanup(func() {
+		goleak.VerifyNone(t)
+	})
+	tests := ListUsersTests{
+		{
+			name: "intersection_panic_expand_rewrite",
+			req: &openfgav1.ListUsersRequest{
+				Object:   &openfgav1.Object{Type: "document", Id: "1"},
+				Relation: "viewer",
+				UserFilters: []*openfgav1.UserTypeFilter{
+					{
+						Type: "user",
+					},
+				},
+			},
+			model: `
+				model
+					schema 1.1
+				type user
+				type document
+					relations
+						define required: [user]
+						define required_other: [user]
+						define viewer: required and required_other`,
+
+			tuples: []*openfgav1.TupleKey{
+				tuple.NewTupleKey("document:1", "required", "user:will"),
+				tuple.NewTupleKey("document:1", "required_other", "user:will"),
+
+				tuple.NewTupleKey("document:1", "required", "user:jon"),
+				tuple.NewTupleKey("document:1", "required_other", "user:maria"),
+			},
+			expectedUsers:     []string{},
+			expectedErrorMsg:  ErrPanic.Error(),
+			newListUsersQuery: NewListUsersQueryPanicExpandIntersectionExpandRewrite,
+		},
+	}
+	tests.runListUsersTestCases(t)
+}
+
 func NewListUsersQueryPanicExpandDirect(ds storage.RelationshipTupleReader, contextualTuples []*openfgav1.TupleKey, opts ...ListUsersQueryOption) *listUsersQuery {
 	l := &listUsersQuery{
 		logger:                  logger.NewNoopLogger(),
@@ -4118,6 +4159,35 @@ func NewListUsersQueryPanicExpandDirect(ds storage.RelationshipTupleReader, cont
 		expandDirectDispatch: func(ctx context.Context, listUsersQuery *listUsersQuery, req *internalListUsersRequest, userObjectType, userObjectID, userRelation string, resp expandResponse, foundUsersChan chan<- foundUser, hasCycle *atomic.Bool) expandResponse {
 			panic(ErrPanic)
 		},
+		expandIntersectionExpandRewrite: expandIntersectionExpandRewrite,
+		expandIntersectionCloseChannels: expandIntersectionCloseChannels,
+		populateFoundUsersCountMap:      populateFoundUsersCountMap,
+	}
+
+	for _, opt := range opts {
+		opt(l)
+	}
+
+	l.datastore = storagewrappers.NewRequestStorageWrapper(ds, contextualTuples, l.maxConcurrentReads)
+
+	return l
+}
+
+func NewListUsersQueryPanicExpandIntersectionExpandRewrite(ds storage.RelationshipTupleReader, contextualTuples []*openfgav1.TupleKey, opts ...ListUsersQueryOption) *listUsersQuery {
+	l := &listUsersQuery{
+		logger:                  logger.NewNoopLogger(),
+		resolveNodeBreadthLimit: serverconfig.DefaultResolveNodeBreadthLimit,
+		resolveNodeLimit:        serverconfig.DefaultResolveNodeLimit,
+		deadline:                serverconfig.DefaultListUsersDeadline,
+		maxResults:              serverconfig.DefaultListUsersMaxResults,
+		maxConcurrentReads:      serverconfig.DefaultMaxConcurrentReadsForListUsers,
+		wasThrottled:            new(atomic.Bool),
+		expandDirectDispatch:    expandDirectDispatch,
+		expandIntersectionExpandRewrite: func(ctx context.Context, l *listUsersQuery, req *internalListUsersRequest, rewrite *openfgav1.Userset, intersectionFoundUsersChans []chan foundUser, i int) expandResponse {
+			panic(ErrPanic)
+		},
+		expandIntersectionCloseChannels: expandIntersectionCloseChannels,
+		populateFoundUsersCountMap:      populateFoundUsersCountMap,
 	}
 
 	for _, opt := range opts {
